fix: Переписать валидацию, переделать мелочовку, пофиксить ошибку при запросах
