+++ errors/unauthorizedError.js
@@ -1,7 +1,6 @@
 class UNAUTHORIZED_ERROR extends Error {
   constructor(message) {
     super(message);
-    this.errorMessage = message;
     this.statusCode = 401;
   }
 }
