+++ middlewares/auth.js
@@ -1,12 +1,11 @@
 const jwt = require('jsonwebtoken');
-const { UNAUTHORIZED_ERROR } = require('../errors/errors');
 const { JWT_SECRET } = require('../config');
 
 function auth(req, res, next) {
   const { authorization } = req.headers;
   if (!authorization || !authorization.startsWith('Bearer ')) {
     return res
-      .status(UNAUTHORIZED_ERROR)
+      .status(401)
       .send({ message: 'Авторизируйтесь, пожалуйста' });
   }
 
@@ -17,7 +16,7 @@ function auth(req, res, next) {
     payload = jwt.verify(token, process.env.JWT_SECRET || JWT_SECRET);
   } catch (err) {
     return res
-      .status(UNAUTHORIZED_ERROR)
+      .status(401)
       .send({ message: 'Авторизируйтесь, пожалуйста' });
   }
