+++ routes/login.js
@@ -1,22 +1,9 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
-const { login, createUser } = require('../controllers/user');
+const { login, createUser, logout } = require('../controllers/user');
+const { loginValidation, createUserValidation } = require('../utils/validation');
 
-const standardValidationTemplate = Joi.string().required();
-
-router.post('/signin', celebrate({
-  body: Joi.object().keys({
-    email: standardValidationTemplate.email(),
-    password: standardValidationTemplate,
-  }),
-}), login);
-
-router.post('/signup', celebrate({
-  body: Joi.object().keys({
-    name: standardValidationTemplate.min(2).max(30),
-    email: standardValidationTemplate.email(),
-    password: standardValidationTemplate,
-  }),
-}), createUser);
+router.post('/signin', loginValidation, login);
+router.post('/signup', createUserValidation, createUser);
+router.post('/signout', logout);
 
 module.exports = router;
